refactor: Reusable horizontal gradient for horizontal lists
