+++ mobile/src/components/Gradient.tsx
@@ -3,11 +3,15 @@
 
 import { LinearGradient } from "expo-linear-gradient";
 import { useMemo } from "react";
+import type { ViewProps } from "react-native";
+import type { AnimatedProps } from "react-native-reanimated";
+import Animated from "react-native-reanimated";
 
 import { cn } from "~/lib/style";
 import type { ColorRole } from "~/modules/customization/theme/core/constants";
 import { useColor } from "~/modules/customization/theme/hooks";
 
+//#region Top Down Gradient
 /** Gradient where the darkest portion is on the top. */
 export function TopDownGradient(props: {
   height: number;
@@ -33,3 +37,48 @@ export function TopDownGradient(props: {
     />
   );
 }
+//#endregion
+
+//#region Horizontal Scroll Gradient
+/** Gradient to smooth out edge transition in horizontal lists. */
+export function HorizontalScrollGradient({
+  children,
+  size = 16,
+  gutter = 0,
+  color,
+  ...props
+}: {
+  children: React.ReactNode;
+  size?: number;
+  gutter?: number;
+  color?: ColorRole;
+} & AnimatedProps<ViewProps>) {
+  const gradientColor = useColor(color, "surface");
+  return (
+    <Animated.View
+      {...props}
+      style={[props.style, { marginHorizontal: -gutter }]}
+      className={cn("relative", props.className)}
+    >
+      {children}
+      {/* Scroll Shadow */}
+      <LinearGradient
+        pointerEvents="none"
+        colors={[`${gradientColor}E6`, `${gradientColor}00`]}
+        {...ShadowProps}
+        style={{ width: size }}
+        className="absolute h-full ltr:left-0 rtl:right-0"
+      />
+      <LinearGradient
+        pointerEvents="none"
+        colors={[`${gradientColor}00`, `${gradientColor}E6`]}
+        {...ShadowProps}
+        style={{ width: size }}
+        className="absolute h-full ltr:right-0 rtl:left-0"
+      />
+    </Animated.View>
+  );
+}
+
+const ShadowProps = { start: { x: 0.0, y: 1.0 }, end: { x: 1.0, y: 1.0 } };
+//#endregion
