+++ mobile/src/navigation/screens/artists/CurrentView.tsx
@@ -3,8 +3,6 @@
 
 import type { StaticScreenProps } from "@react-navigation/native";
 import { useNavigation } from "@react-navigation/native";
-import { LinearGradient } from "expo-linear-gradient";
-import { View } from "react-native";
 
 import { useArtistDetails, useArtistTracks } from "~/data/artist/queries";
 import type { ArtistAlbum } from "~/data/artist/types";
@@ -18,10 +16,10 @@ import { CurrentListMenu } from "~/navigation/components/CurrentListMenu";
 import { PagePlaceholder } from "~/navigation/components/Placeholder";
 
 import { FlatList } from "~/components/Base/List";
+import { HorizontalScrollGradient } from "~/components/Gradient";
 import { SafeContainer } from "~/components/SafeContainer";
 import { useSheetRef } from "~/components/Sheet/useSheetRef";
 import { TEm } from "~/components/Typography/StyledText";
-import { useTheme } from "~/modules/customization/theme/hooks";
 import { MediaCard } from "~/modules/media/components/MediaCard";
 import { useTrackListPreset } from "~/modules/media/components/Track";
 
@@ -90,14 +88,13 @@ export default function Artist({
  */
 function ArtistAlbums({ albums }: { albums: ArtistAlbum[] | null }) {
   const navigation = useNavigation();
-  const { surface } = useTheme();
   const { width } = useGetColumn(ColumnPresets.horizontalList);
 
   if (!albums) return null;
   return (
     <>
       <TEm textKey="term.albums" className="mb-2" />
-      <View className="relative -mx-4">
+      <HorizontalScrollGradient gutter={16}>
         <FlatList
           horizontal
           data={albums}
@@ -118,24 +115,8 @@ function ArtistAlbums({ albums }: { albums: ArtistAlbum[] | null }) {
           )}
           contentContainerClassName="px-4"
         />
-
-        {/* Scroll Shadow */}
-        <LinearGradient
-          pointerEvents="none"
-          colors={[`${surface}E6`, `${surface}00`]}
-          {...ShadowProps}
-          className="absolute h-full w-4 ltr:left-0 rtl:right-0"
-        />
-        <LinearGradient
-          pointerEvents="none"
-          colors={[`${surface}00`, `${surface}E6`]}
-          {...ShadowProps}
-          className="absolute h-full w-4 ltr:right-0 rtl:left-0"
-        />
-      </View>
+      </HorizontalScrollGradient>
       <TEm textKey="term.tracks" className="mt-4 mb-2" />
     </>
   );
 }
-
-const ShadowProps = { start: { x: 0.0, y: 1.0 }, end: { x: 1.0, y: 1.0 } };
